Improve the logic of RocksUniversalCompaction
Summary:
RocksUniversalCompaction now proceeds as follows:
1. Find the first level which has files
2. Compute the estimated output size.
3. If the output size exceed the size limit of the current level,
   then include files in the current level; level++ and repeat
   step 2 and 3 until the current level can fit the output.

Size limit of each level is determined by the following equation, which is
the same as level compaction.
    max_bytes_for_level_base * max_bytes_for_level_multiplier^level

When max_bytes_for_level_multiplier is set to 2, then it is equivalent to
universal compaction.

This patch also fixed the errors where number of compaction input is hard
coded to 2 (while D20577 is waiting for approval in the master)

Next Step: fix the doubling-in-space issue of universal compaction.

Test Plan: Now all test passed!!

Reviewers: igor, ljin, sdong

Subscribers: leveldb

Differential Revision: https://reviews.facebook.net/D20583